requirements.txt
jpegio
cython
pillow
pandas
numpy
streamlit
wx
